requirements.txt
dnspython>=2.7.0
pytest>=7.0.0
pytest-cov>=4.0.0